only output colours if not `--machine-readable`
