+++ source/class/qx/tool/cli/Cli.js
@@ -18,7 +18,6 @@
 
 require("@qooxdoo/framework");
 const path = require("path");
-const consoleControl = require("console-control-strings");
 
 /**
  * Entry point for the CLI
@@ -30,17 +29,6 @@ qx.Class.define("qx.tool.cli.Cli", {
 
   members: {
     async run() {
-      let configDb = await qx.tool.cli.ConfigDb.getInstance();
-      let color = configDb.db("qx.default.color", null);
-      if (color) {
-        let colorOn = consoleControl.color(color.split(" "));
-        process.stdout.write(colorOn + consoleControl.eraseLine());
-        let colorReset = consoleControl.color("reset");
-        process.on("exit", () => process.stdout.write(colorReset + consoleControl.eraseLine()));
-        let Console = qx.tool.compiler.Console.getInstance();
-        Console.setColorOn(colorOn);
-      }
-      
       var args = qx.lang.Array.clone(process.argv);
       args.shift();
       process.title = args.join(" ");
